Revert "moved the stateless functions to session utils module"
This reverts commit 17e04e51950731aa29e2ca32fc8294e0cc45eb98.
