+++ crates/goose-cli/src/session/mod.rs
@@ -5,7 +5,6 @@ mod input;
 mod output;
 mod prompt;
 mod thinking;
-mod utils;
 
 pub use self::export::message_to_markdown;
 pub use builder::{build_session, SessionBuilderConfig, SessionSettings};
@@ -73,10 +72,41 @@ impl CompletionCache {
     }
 }
 
-// Re-export types and functions from utils module for backward compatibility
-pub use utils::{
-    classify_planner_response, get_reasoner, summarize_context_messages, PlannerResponseType,
-};
+pub enum PlannerResponseType {
+    Plan,
+    ClarifyingQuestions,
+}
+
+/// Decide if the planner's reponse is a plan or a clarifying question
+///
+/// This function is called after the planner has generated a response
+/// to the user's message. The response is either a plan or a clarifying
+/// question.
+pub async fn classify_planner_response(
+    message_text: String,
+    provider: Arc<dyn Provider>,
+) -> Result<PlannerResponseType> {
+    let prompt = format!("The text below is the output from an AI model which can either provide a plan or list of clarifying questions. Based on the text below, decide if the output is a \"plan\" or \"clarifying questions\".\n---\n{message_text}");
+
+    // Generate the description
+    let message = Message::user().with_text(&prompt);
+    let (result, _usage) = provider
+        .complete(
+            "Reply only with the classification label: \"plan\" or \"clarifying questions\"",
+            &[message],
+            &[],
+        )
+        .await?;
+
+    // println!("classify_planner_response: {result:?}\n"); // TODO: remove
+
+    let predicted = result.as_concat_text();
+    if predicted.to_lowercase().contains("plan") {
+        Ok(PlannerResponseType::Plan)
+    } else {
+        Ok(PlannerResponseType::ClarifyingQuestions)
+    }
+}
 
 impl Session {
     pub fn new(
@@ -109,6 +139,21 @@ impl Session {
             scheduled_job_id,
             max_turns,
         }
+    }
+
+    /// Helper function to summarize context messages
+    async fn summarize_context_messages(
+        messages: &mut Vec<Message>,
+        agent: &Agent,
+        message_suffix: &str,
+    ) -> Result<()> {
+        // Summarize messages to fit within context length
+        let (summarized_messages, _) = agent.summarize_context(messages).await?;
+        let msg = format!("Context maxed out\n{}\n{}", "-".repeat(50), message_suffix);
+        output::render_text(&msg, Some(Color::Yellow), true);
+        *messages = summarized_messages;
+
+        Ok(())
     }
 
     /// Add a stdio extension to the session
@@ -845,7 +890,7 @@ impl Session {
                                         } else {
                                             "Goose automatically summarized messages to continue processing."
                                         };
-                                        summarize_context_messages(&mut self.messages, &self.agent, message_suffix).await?;
+                                        Self::summarize_context_messages(&mut self.messages, &self.agent, message_suffix).await?;
                                     }
                                     _ => {
                                         unreachable!()
@@ -1355,4 +1400,36 @@ impl Session {
 
         Ok(path)
     }
+}
+
+fn get_reasoner() -> Result<Arc<dyn Provider>, anyhow::Error> {
+    use goose::model::ModelConfig;
+    use goose::providers::create;
+
+    let config = Config::global();
+
+    // Try planner-specific provider first, fallback to default provider
+    let provider = if let Ok(provider) = config.get_param::<String>("GOOSE_PLANNER_PROVIDER") {
+        provider
+    } else {
+        println!("WARNING: GOOSE_PLANNER_PROVIDER not found. Using default provider...");
+        config
+            .get_param::<String>("GOOSE_PROVIDER")
+            .expect("No provider configured. Run 'goose configure' first")
+    };
+
+    // Try planner-specific model first, fallback to default model
+    let model = if let Ok(model) = config.get_param::<String>("GOOSE_PLANNER_MODEL") {
+        model
+    } else {
+        println!("WARNING: GOOSE_PLANNER_MODEL not found. Using default model...");
+        config
+            .get_param::<String>("GOOSE_MODEL")
+            .expect("No model configured. Run 'goose configure' first")
+    };
+
+    let model_config = ModelConfig::new(model);
+    let reasoner = create(&provider, model_config)?;
+
+    Ok(reasoner)
 }
